requirements.txt
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
djangorestframework==3.12.2
idna==2.8
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
urllib3==1.25.6
wcwidth==0.1.8
zipp==2.2.0
